Dummy Recommendations

+++ database/migrations/007_create_recommendations.rb
@@ -4,7 +4,7 @@ Sequel.migration do
       primary_key :id, index: true
       foreign_key :user_id, :users
       foreign_key :print_id, :prints
-      Integer :rating
+      Float :rating
       String :comment
     end
   end
